requirements.txt
Flask==2.3.2
Flask_Login==0.6.2
pytest==7.4.0
python_bcrypt==0.3.2
Requests==2.31.0
